[general] changed dcxDrawRect() to call RoundRect() with a width & height of 5 (this may chnage again if it looks odd somewhere) changed dcxDrawBitMap() to take an bAlpha arg to enabled blending the bitmap. added some sanity checks to FontDialog refactored CreateBMPFile() moved some local vars to an anonymous namespace.
[dialog]
changed /xdialog -p to support enabling shadow text via +s flag

[xpopup]
changed g_menuTheme to not be static.
renamed g_menuTheme as m_menuTheme
added m_bDrawShadowText setting.
updated custom menubar code to support changing border colours.

+++ DxErr/dxerr.cpp
@@ -3467,7 +3467,8 @@ void WINAPI DXGetErrorDescriptionW(_In_ HRESULT hr, _Out_cap_(count) WCHAR* desc
 	*desc = 0;
 
 	// First try to see if FormatMessage knows this hr
-	const UINT icount = static_cast<UINT>(std::min<size_t>(count, 32767U));
+	//const UINT icount = static_cast<UINT>(std::min<size_t>(count, 32767U));
+	const DWORD icount = std::min<DWORD>(static_cast<DWORD>(count), 32767U);
 
 	const DWORD result = FormatMessageW(FORMAT_MESSAGE_FROM_SYSTEM, nullptr, hr,
 		MAKELANGID(LANG_NEUTRAL, SUBLANG_DEFAULT), desc, icount, nullptr);
